+++ code/hplus_instance.cpp
@@ -52,6 +52,7 @@ void hplus::init(environment& env) {
                       .tight_bounds = false,
                       .minimal_landmark = true,
                       .complete_landmark = true,
+                      .sec = true,
                       .using_cplex = true,
                       .log = false,
                       .write_lp = false,